Make the list use common or scientific name instead of and
